workshop_cv_day2/demo/backend/requirements.txt
fastapi
uvicorn
python-multipart
deepface
Pillow
numpy
tf-keras
